+++ src/AdminSystemServiceProvider.php
@@ -59,6 +59,8 @@ class AdminSystemServiceProvider extends ServiceProvider
 
     public function register()
     {
+        $this->registerServiceProviders();
+
         $this->mergeConfigFrom(
             __DIR__ . '/../config/admin.php',
             'admin'
@@ -85,6 +87,12 @@ class AdminSystemServiceProvider extends ServiceProvider
         );
 
         $this->registerServices();
+    }
+
+    protected function registerServiceProviders()
+    {
+        $this->app->register(\Tymon\JWTAuth\Providers\LaravelServiceProvider::class);
+        $this->app->register(\L5Swagger\L5SwaggerServiceProvider::class);
     }
 
     protected function registerServices()