+++ src/Middleware/JwtMiddleware.php
@@ -5,6 +5,7 @@ namespace Cheney\AdminSystem\Middleware;
 use Closure;
 use Illuminate\Http\Request;
 use Cheney\AdminSystem\Traits\ApiResponseTrait;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class JwtMiddleware
 {
@@ -13,7 +14,7 @@ class JwtMiddleware
     public function handle(Request $request, Closure $next)
     {
         try {
-            $admin = auth('admin')->user();
+            $admin = JWTAuth::parseToken()->authenticate();
             
             if (!$admin) {
                 return $this->unauthorized();